Refactor utils into a muliple modules
