+++ sail_on_client/protocol/ond_round.py
@@ -11,7 +11,8 @@ from sail_on_client.protocol.ond_dataclasses import (
 )
 from sail_on_client.protocol.parinterface import ParInterface
 from sail_on_client.protocol.localinterface import LocalInterface
-from sail_on_client.utils import skip_stage, safe_remove
+from sail_on_client.utils.utils import safe_remove
+from sail_on_client.utils.decorators import skip_stage
 
 
 log = logging.getLogger(__name__)
