+++ sail_on_client/utils/__init__.py
@@ -1,0 +1,1 @@
+"""Sail On client utility package."""
